archive/python_original/requirements.txt
# Core dependencies for OHdio Audiobook Downloader
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Async HTTP
aiohttp>=3.9.0

# Media download
yt-dlp>=2023.12.30

# Audio metadata
mutagen>=1.47.0

# Image processing
Pillow>=9.0.0

# Progress bars and UI
tqdm>=4.64.0
colorama>=0.4.6
gradio>=4.44.0
